ExploreDesktop as own component, pass icon Prop to Dropdown search
