+++ src/components/Navigation/DesktopNavigation/index.js
@@ -1,67 +1,20 @@
-import { Grid, Typography } from "@material-ui/core";
+import { Grid } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import PropTypes from "prop-types";
 import React from "react";
 
-import DropdownSearch from "@/pesayetu/components/DropdownSearch";
 import Logo from "@/pesayetu/components/Logo";
 import Menu from "@/pesayetu/components/Menu";
 import Section from "@/pesayetu/components/Section";
 
-const useStyles = makeStyles(({ palette, typography }) => ({
+const useStyles = makeStyles(() => ({
   root: {},
   section: {},
-  help: {
-    color: "#666666",
-    textAlign: "center",
-    backgroundColor: "#EBEBEB",
-    borderRadius: typography.pxToRem(60),
-    marginLeft: typography.pxToRem(20),
-    width: typography.pxToRem(48),
-    height: typography.pxToRem(48),
-  },
-  searchLabel: {
-    display: "none",
-  },
-  searchInput: {
-    borderRadius: 0,
-    padding: `0 ${typography.pxToRem(10)}`,
-    color: "#959696",
-    textTransform: "initial",
-    "&::placeholder": {
-      opacity: 1,
-    },
-  },
-  searchInputRoot: {
-    borderRadius: 0,
-    backgroundColor: palette.background.paper,
-    borderColor: palette.background.default,
-  },
-  selectMenu: {
-    borderRadius: 0,
-    border: 0,
-    background: palette.background.paper,
-    marginTop: typography.pxToRem(2),
-  },
-  searchMenuItem: {
-    "&:hover": {
-      color: palette.text.secondary,
-      background: palette.primary.main,
-    },
-  },
 }));
 
-function DesktopNavigation({
-  isExplore,
-  logoProps,
-  menuProps,
-  socialLinks,
-  ...props
-}) {
+function DesktopNavigation({ logoProps, menuProps, socialLinks, ...props }) {
   const classes = useStyles(props);
-  const openTooltip = () => {
-    // tooltip
-  };
+
   return (
     <div className={classes.root}>
       <Section classes={{ root: classes.section }}>
@@ -77,31 +30,7 @@ function DesktopNavigation({
             justifyContent="flex-end"
             alignItems="center"
           >
-            {isExplore ? (
-              <>
-                <DropdownSearch
-                  isExplore
-                  {...props}
-                  classes={{
-                    inputRoot: classes.searchInputRoot,
-                    input: classes.searchInput,
-                    label: classes.searchLabel,
-                    selectMenu: classes.selectMenu,
-                    menuItem: classes.searchMenuItem,
-                  }}
-                />
-                <Typography
-                  component="div"
-                  onClick={openTooltip}
-                  variant="h3"
-                  className={classes.help}
-                >
-                  ?
-                </Typography>
-              </>
-            ) : (
-              <Menu links={menuProps} socialLinks={socialLinks} />
-            )}
+            <Menu links={menuProps} socialLinks={socialLinks} />
           </Grid>
           <Grid />
         </Grid>
@@ -114,14 +43,12 @@ DesktopNavigation.propTypes = {
   logoProps: PropTypes.shape({}),
   menuProps: PropTypes.arrayOf(PropTypes.shape({})),
   socialLinks: PropTypes.arrayOf(PropTypes.shape({})),
-  isExplore: PropTypes.bool,
 };
 
 DesktopNavigation.defaultProps = {
   logoProps: undefined,
   menuProps: undefined,
   socialLinks: undefined,
-  isExplore: false,
 };
 
 export default DesktopNavigation;
